+++ DynamicGameAssets/Game/CustomBasicFurniture.cs
@@ -132,16 +132,20 @@ namespace DynamicGameAssets.Game
 
             spriteBatch.Draw(currTex.Texture, location, currTex.Rect, Color.White * alpha, 0f, Vector2.Zero, 4f, this.flipped ? SpriteEffects.FlipHorizontally : SpriteEffects.None, layerDepth);
         }
-        
+
+        /// <inheritdoc />
         public override void AddLightGlow(GameLocation location)
         {
+            // equivalent to the base logic, but correct lighting position for custom furniture
             if (!this.lightGlowPosition.HasValue)
             {
-                Vector2 vector = new Vector2((int)base.tileLocation.X * 64 + 32, (int)base.tileLocation.Y * 64 + 64);
-                if (!location.lightGlows.Contains(vector))
+                Vector2 furniturePixel = this.TileLocation * Game1.tileSize;
+                Vector2 glowPixel = furniturePixel + new Vector2((float)Game1.tileSize / 2, Game1.tileSize);
+
+                if (!location.lightGlows.Contains(glowPixel))
                 {
-                    this.lightGlowPosition = vector;
-                    location.lightGlows.Add(vector);
+                    this.lightGlowPosition = glowPixel;
+                    location.lightGlows.Add(glowPixel);
                 }
             }
         }